requirements.txt
anyio==4.3.0
certifi==2024.2.2
click==8.1.2
colorama==0.4.4
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.26.0
idna==3.6
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.1
python-telegram-bot==20.8
sniffio==1.3.0
tomli==2.0.1
typing_extensions==4.9.0
